Use npx for launching ts-node
