+++ package.json
@@ -10,8 +10,6 @@
     "format": "biome format . --write",
     "prebuild": "npm run lint",
     "build": "tsc",
-    "test:coverage": "jest --verbose --coverage --detectOpenHandles --forceExit",
-    "codecov": "codecov",
     "test": "jest --verbose --detectOpenHandles --forceExit",
     "gen-proto": "sh genproto.sh",
     "prepublishOnly": "tsc"
@@ -20,7 +18,12 @@
     "type": "git",
     "url": "git+https://github.com/bugdiver/gauge-ts.git"
   },
-  "keywords": ["gauge", "Acceptance Test", "Typescript", "Automation"],
+  "keywords": [
+    "gauge",
+    "Acceptance Test",
+    "Typescript",
+    "Automation"
+  ],
   "author": "BugDiver <vinayshankar00@gmail.com>",
   "license": "MIT",
   "bugs": {
@@ -30,7 +33,7 @@
     "@grpc/grpc-js": "^1.10.8",
     "google-protobuf": "^3.13.0",
     "klaw-sync": "^6.0.0",
-    "ts-node": "8.7.0",
+    "ts-node": "^10.9.0",
     "typescript": "^5.0.4",
     "uuid": "^8.3.0"
   },
@@ -40,9 +43,6 @@
     "@types/jest": "^24.9.1",
     "@types/klaw-sync": "^6.0.0",
     "@types/uuid": "^8.3.0",
-    "@typescript-eslint/eslint-plugin": "^3.9.0",
-    "@typescript-eslint/parser": "^3.9.0",
-    "codecov": "^3.7.2",
     "grpc_tools_node_protoc_ts": "^5.3.3",
     "grpc-tools": "^1.12.4",
     "jest": "^29.7.0",
