+++ src/GaugeRuntime.ts
@@ -18,7 +18,6 @@ export class GaugeRuntime {
           throw err;
         }
         console.log(`Listening on port:${port}`);
-        server.start();
       },
     );
   }
