+++ tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
-    "target": "es6",
+    "target": "esnext",
     "module": "commonjs",
     "lib": [
-      "es2016"
+      "es2023"
     ],
     "declaration": true,
     "outDir": "./dist",
